+++ extension/src/inpage/middleware/createRpcWarningMiddleware.ts
