--- extension/src/inpage/provider/BaseProvider.ts
@@ -1,414 +1,0 @@
-import { EthereumRpcError, ethErrors } from "eth-rpc-errors";
-import dequal from "fast-deep-equal";
-import {
-  JsonRpcEngine,
-  JsonRpcId,
-  JsonRpcMiddleware,
-  JsonRpcRequest,
-  JsonRpcVersion,
-} from "json-rpc-engine";
-import log from "loglevel";
-
-import SafeEventEmitter from "@metamask/safe-event-emitter";
-
-import messages from "./messages";
-import { Maybe, getRpcPromiseCallback, isValidChainId } from "./utils";
-
-export interface UnvalidatedJsonRpcRequest {
-  id?: JsonRpcId;
-  jsonrpc?: JsonRpcVersion;
-  method: string;
-  params?: unknown;
-}
-
-export interface BaseProviderOptions {
-  /**
-   * The maximum number of event listeners.
-   */
-  maxEventListeners?: number;
-
-  /**
-   * `json-rpc-engine` middleware. The middleware will be inserted in the given
-   * order immediately after engine initialization.
-   */
-  rpcMiddleware?: JsonRpcMiddleware<unknown, unknown>[];
-}
-
-export interface RequestArguments {
-  /** The RPC method to request. */
-  method: string;
-
-  /** The params of the RPC method, if any. */
-  params?: unknown[] | Record<string, unknown>;
-}
-
-export interface BaseProviderState {
-  accounts: null | string[];
-  isConnected: boolean;
-  isUnlocked: boolean;
-  initialized: boolean;
-  isPermanentlyDisconnected: boolean;
-}
-
-/**
- * An abstract class implementing the EIP-1193 interface. Implementers must:
- *
- * 1. At initialization, push a middleware to the internal `_rpcEngine` that
- *    hands off requests to the server and receives responses in return.
- * 2. At initialization, retrieve initial state and call
- *    {@link BaseProvider._initializeState} **once**.
- * 3. Ensure that the provider's state is synchronized with the wallet.
- * 4. Ensure that notifications are received and emitted as appropriate.
- */
-export abstract class BaseProvider extends SafeEventEmitter {
-  protected _state: BaseProviderState;
-
-  protected _rpcEngine: JsonRpcEngine;
-
-  protected static _defaultState: BaseProviderState = {
-    accounts: null,
-    isConnected: false,
-    isUnlocked: false,
-    initialized: false,
-    isPermanentlyDisconnected: false,
-  };
-
-  /**
-   * The chain ID of the currently connected Ethereum chain.
-   * See [chainId.network]{@link https://chainid.network} for more information.
-   */
-  public chainId: string | null;
-
-  /**
-   * The user's currently selected Ethereum address.
-   * If null, Iron is either locked or the user has not permitted any
-   * addresses to be viewed.
-   */
-  public selectedAddress: string | null;
-
-  /**
-   * @param options - An options bag
-   * @param options.logger - The logging API to use. Default: console
-   * @param options.maxEventListeners - The maximum number of event
-   * listeners. Default: 100
-   */
-  constructor({
-    maxEventListeners = 100,
-    rpcMiddleware = [],
-  }: BaseProviderOptions = {}) {
-    super();
-
-    this.setMaxListeners(maxEventListeners);
-
-    // Private state
-    this._state = {
-      ...BaseProvider._defaultState,
-    };
-
-    // Public state
-    this.selectedAddress = null;
-    this.chainId = null;
-
-    // Bind functions to prevent consumers from making unbound calls
-    this._handleAccountsChanged = this._handleAccountsChanged.bind(this);
-    this._handleConnect = this._handleConnect.bind(this);
-    this._handleChainChanged = this._handleChainChanged.bind(this);
-    this._handleDisconnect = this._handleDisconnect.bind(this);
-    this._handleUnlockStateChanged = this._handleUnlockStateChanged.bind(this);
-    this._rpcRequest = this._rpcRequest.bind(this);
-    this.request = this.request.bind(this);
-
-    // Handle RPC requests via dapp-side RPC engine.
-    //
-    // ATTN: Implementers must push a middleware that hands off requests to
-    // the server.
-    const rpcEngine = new JsonRpcEngine();
-    rpcMiddleware.forEach((middleware) => rpcEngine.push(middleware));
-    this._rpcEngine = rpcEngine;
-  }
-
-  //====================
-  // Public Methods
-  //====================
-
-  /**
-   * Returns whether the provider can process RPC requests.
-   */
-  isConnected(): boolean {
-    return this._state.isConnected;
-  }
-
-  /**
-   * Submits an RPC request for the given method, with the given params.
-   * Resolves with the result of the method call, or rejects on error.
-   *
-   * @param args - The RPC request arguments.
-   * @param args.method - The RPC method name.
-   * @param args.params - The parameters for the RPC method.
-   * @returns A Promise that resolves with the result of the RPC method,
-   * or rejects if an error is encountered.
-   */
-  async request<T>(args: RequestArguments): Promise<Maybe<T>> {
-    if (!args || typeof args !== "object" || Array.isArray(args)) {
-      throw ethErrors.rpc.invalidRequest({
-        message: messages.errors.invalidRequestArgs(),
-        data: args,
-      });
-    }
-
-    const { method, params } = args;
-
-    if (typeof method !== "string" || method.length === 0) {
-      throw ethErrors.rpc.invalidRequest({
-        message: messages.errors.invalidRequestMethod(),
-        data: args,
-      });
-    }
-
-    if (
-      params !== undefined &&
-      !Array.isArray(params) &&
-      (typeof params !== "object" || params === null)
-    ) {
-      throw ethErrors.rpc.invalidRequest({
-        message: messages.errors.invalidRequestParams(),
-        data: args,
-      });
-    }
-
-    return new Promise<T>((resolve, reject) => {
-      this._rpcRequest(
-        { method, params },
-        getRpcPromiseCallback(resolve as any, reject) as any
-      );
-    });
-  }
-
-  //====================
-  // Private Methods
-  //====================
-
-  /**
-   * **MUST** be called by child classes.
-   *
-   * Sets initial state if provided and marks this provider as initialized.
-   * Throws if called more than once.
-   *
-   * Permits the `networkVersion` field in the parameter object for
-   * compatibility with child classes that use this value.
-   *
-   * @param initialState - The provider's initial state.
-   * @emits BaseProvider#_initialized
-   * @emits BaseProvider#connect - If `initialState` is defined.
-   */
-  protected _initializeState(initialState?: {
-    accounts: string[];
-    chainId: string;
-    isUnlocked: boolean;
-    networkVersion?: string;
-  }) {
-    if (this._state.initialized === true) {
-      throw new Error("Provider already initialized.");
-    }
-
-    if (initialState) {
-      const { accounts, chainId, isUnlocked, networkVersion } = initialState;
-
-      // EIP-1193 connect
-      this._handleConnect(chainId);
-      this._handleChainChanged({ chainId, networkVersion });
-      this._handleUnlockStateChanged({ accounts, isUnlocked });
-      this._handleAccountsChanged(accounts);
-    }
-
-    // Mark provider as initialized regardless of whether initial state was
-    // retrieved.
-    this._state.initialized = true;
-    this.emit("_initialized");
-  }
-
-  /**
-   * Internal RPC method. Forwards requests to background via the RPC engine.
-   * Also remap ids inbound and outbound.
-   *
-   * @param payload - The RPC request object.
-   * @param callback - The consumer's callback.
-   */
-  protected _rpcRequest(
-    payload: UnvalidatedJsonRpcRequest | UnvalidatedJsonRpcRequest[],
-    cb: (...args: unknown[]) => void
-  ) {
-    if (!Array.isArray(payload)) {
-      if (!payload.jsonrpc) {
-        payload.jsonrpc = "2.0";
-      }
-
-      return this._rpcEngine.handle(payload as JsonRpcRequest<unknown>, cb);
-    }
-    return this._rpcEngine.handle(payload as JsonRpcRequest<unknown>[], cb);
-  }
-
-  /**
-   * When the provider becomes connected, updates internal state and emits
-   * required events. Idempotent.
-   *
-   * @param chainId - The ID of the newly connected chain.
-   * @emits IronInpageProvider#connect
-   */
-  protected _handleConnect(chainId: string) {
-    if (!this._state.isConnected) {
-      this._state.isConnected = true;
-      this.emit("connect", { chainId });
-      log.debug(messages.info.connected(chainId));
-    }
-  }
-
-  /**
-   * When the provider becomes disconnected, updates internal state and emits
-   * required events. Idempotent with respect to the isRecoverable parameter.
-   *
-   * Error codes per the CloseEvent status codes as required by EIP-1193:
-   * https://developer.mozilla.org/en-US/docs/Web/API/CloseEvent#Status_codes
-   *
-   * @param isRecoverable - Whether the disconnection is recoverable.
-   * @param errorMessage - A custom error message.
-   * @emits BaseProvider#disconnect
-   */
-  protected _handleDisconnect(isRecoverable: boolean, errorMessage?: string) {
-    if (
-      this._state.isConnected ||
-      (!this._state.isPermanentlyDisconnected && !isRecoverable)
-    ) {
-      this._state.isConnected = false;
-
-      let error;
-      if (isRecoverable) {
-        error = new EthereumRpcError(
-          1013, // Try again later
-          errorMessage || messages.errors.disconnected()
-        );
-        log.debug(error);
-      } else {
-        error = new EthereumRpcError(
-          1011, // Internal error
-          errorMessage || messages.errors.permanentlyDisconnected()
-        );
-
-        log.error(error);
-        this.chainId = null;
-        this._state.accounts = null;
-        this.selectedAddress = null;
-        this._state.isUnlocked = false;
-        this._state.isPermanentlyDisconnected = true;
-      }
-
-      this.emit("disconnect", error);
-    }
-  }
-
-  /**
-   * Upon receipt of a new `chainId`, emits the corresponding event and sets
-   * and sets relevant public state. Does nothing if the given `chainId` is
-   * equivalent to the existing value.
-   *
-   * Permits the `networkVersion` field in the parameter object for
-   * compatibility with child classes that use this value.
-   *
-   * @emits BaseProvider#chainChanged
-   * @param networkInfo - An object with network info.
-   * @param networkInfo.chainId - The latest chain ID.
-   */
-  protected _handleChainChanged({
-    chainId,
-  }: { chainId?: string; networkVersion?: string } = {}) {
-    if (!isValidChainId(chainId)) {
-      log.error(messages.errors.invalidNetworkParams(), { chainId });
-      return;
-    }
-
-    this._handleConnect(chainId);
-
-    if (chainId !== this.chainId) {
-      this.chainId = chainId;
-      if (this._state.initialized) {
-        this.emit("chainChanged", this.chainId);
-      }
-    }
-  }
-
-  /**
-   * Called when accounts may have changed. Diffs the new accounts value with
-   * the current one, updates all state as necessary, and emits the
-   * accountsChanged event.
-   *
-   * @param accounts - The new accounts value.
-   * @param isEthAccounts - Whether the accounts value was returned by
-   * a call to eth_accounts.
-   */
-  protected _handleAccountsChanged(accounts: unknown[]): void {
-    let _accounts = accounts;
-
-    if (!Array.isArray(accounts)) {
-      log.error(
-        "Iron: Received invalid accounts parameter. Please report this bug.",
-        accounts
-      );
-      _accounts = [];
-    }
-
-    for (const account of accounts) {
-      if (typeof account !== "string") {
-        log.error(
-          "Iron: Received non-string account. Please report this bug.",
-          accounts
-        );
-        _accounts = [];
-        break;
-      }
-    }
-
-    // emit accountsChanged if anything about the accounts array has changed
-    if (!dequal(this._state.accounts, _accounts)) {
-      this._state.accounts = _accounts as string[];
-
-      // handle selectedAddress
-      if (this.selectedAddress !== _accounts[0]) {
-        this.selectedAddress = (_accounts[0] as string) || null;
-      }
-
-      // finally, after all state has been updated, emit the event
-      if (this._state.initialized) {
-        this.emit("accountsChanged", _accounts);
-      }
-    }
-  }
-
-  /**
-   * Upon receipt of a new isUnlocked state, sets relevant public state.
-   * Calls the accounts changed handler with the received accounts, or an empty
-   * array.
-   *
-   * Does nothing if the received value is equal to the existing value.
-   * There are no lock/unlock events.
-   *
-   * @param opts - Options bag.
-   * @param opts.accounts - The exposed accounts, if any.
-   * @param opts.isUnlocked - The latest isUnlocked value.
-   */
-  protected _handleUnlockStateChanged({
-    accounts,
-    isUnlocked,
-  }: { accounts?: string[]; isUnlocked?: boolean } = {}) {
-    if (typeof isUnlocked !== "boolean") {
-      log.error(
-        "Iron: Received invalid isUnlocked parameter. Please report this bug."
-      );
-      return;
-    }
-
-    if (isUnlocked !== this._state.isUnlocked) {
-      this._state.isUnlocked = isUnlocked;
-      this._handleAccountsChanged(accounts || []);
-    }
-  }
-}
