--- extension/src/inpage/provider/index.ts
@@ -1,1 +1,0 @@
-export { initializeProvider } from "./initializeInpageProvider";
