+++ extension/src/inpage/utils.ts
